% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_dados.R
\encoding{UTF-8}
\name{atualizar_dados}
\alias{atualizar_dados}
\title{Atualiza os dados de um SpatialPolygonsDataFrame com base em uma expressão fornecida}
\usage{
atualizar_dados(sp_df, expressao)
}
\arguments{
\item{sp_df}{Um objeto `SpatialPolygonsDataFrame` que será atualizado.}

\item{expressao}{Uma expressão que será avaliada usando a função `obter_dados`.}
}
\value{
Um `SpatialPolygonsDataFrame` atualizado com os dados resultantes da expressão fornecida.
}
\description{
Esta função permite que os usuários atualizem os dados associados a um \code{SpatialPolygonsDataFrame} usando uma expressão que referencia tabelas e variáveis específicas. A função utiliza a expressão fornecida para obter os novos dados e atualiza o \code{SpatialPolygonsDataFrame} com esses dados.
Atualizar Dados em um SpatialPolygonsDataFrame
}
\details{
Esta função atualiza a coluna 'dados' de um `SpatialPolygonsDataFrame` com base em uma expressão fornecida.
A expressão é avaliada usando a função `obter_dados` e os resultados são mapeados para o `SpatialPolygonsDataFrame` com base na coluna 'Name'.


As variáveis que podem ser referenciadas na expressão, bem como detalhes sobre elas,
estão disponíveis no data.frame "dicionario" incluído no pacote. Este data.frame fornece
uma descrição detalhada de cada variável disponível.
}
\examples{
# Suponha que sp_df seja um SpatialPolygonsDataFrame e expressao uma expressão válida:
# atualizar_dados(sp_df, expressao)

}
