% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{dicionario}
\alias{dicionario}
\title{Dicionário de Variáveis}
\format{
Um data frame com 6 linhas e 2 colunas:
\describe{
  \item{Variável}{Caracter. Nome da variável. Exemplo: "1.V001".}
  \item{Descrição}{Caracter. Descrição detalhada da variável. Exemplo: "Domicílios particulares permanentes ou pessoas responsáveis por domicílios particulares permanentes".}
}
}
\source{
IBGE, Censo 2010.
}
\usage{
dicionario
}
\description{
Este data frame contém a descrição das variáveis utilizadas no conjunto de dados.
}
