% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exibir_mapa.R
\name{exibir_mapa}
\alias{exibir_mapa}
\title{Exibir Mapa com Dados do Censo}
\usage{
exibir_mapa(mapa)
}
\arguments{
\item{mapa}{Um objeto SpatialPolygonsDataFrame contendo os dados do censo a serem exibidos.}
}
\value{
Um objeto leaflet para visualização interativa.
}
\description{
Esta função exibe um mapa interativo com base nos dados do censo fornecidos.
Se a coluna 'dados' estiver presente no mapa, os polígonos serão coloridos de acordo com esses dados.
Caso contrário, os polígonos serão exibidos com rótulos baseados na coluna 'Description'.
}
\examples{
# Suponha que 'mapa' seja um SpatialPolygonsDataFrame válido:
# exibir_mapa(mapa)

}
