% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_da_cidade.R
\encoding{UTF-8}
\name{mapa_da_cidade}
\alias{mapa_da_cidade}
\title{Obter Mapa da Cidade com Dados do Censo}
\usage{
mapa_da_cidade(
  cidade = "VITÓRIA",
  dados = NULL,
  ano = 2010,
  eliminar_inabitadas = FALSE,
  eliminar_agua = FALSE,
  eliminar_sobrepostas = FALSE
)
}
\arguments{
\item{cidade}{Nome da cidade para a qual o mapa deve ser obtido. O padrão é "VITÓRIA".}

\item{dados}{Expressão ou coluna de dados a serem adicionados ao mapa.}

\item{ano}{Ano do censo. Atualmente, apenas 2010 é suportado.}

\item{eliminar_inabitadas}{Se TRUE, elimina regiões inabitadas do mapa.}

\item{eliminar_agua}{Se TRUE, elimina massas de água do mapa.}

\item{eliminar_sobrepostas}{Se TRUE, elimina áreas sobrepostas no mapa.}
}
\value{
Um objeto SpatialPolygonsDataFrame contendo o mapa da cidade com os dados do censo.
}
\description{
Esta função retorna um mapa da cidade especificada com base nos dados do censo do IBGE.
O usuário pode optar por eliminar regiões inabitadas, massas de água e áreas sobrepostas.
}
\examples{
# Suponha que 'dados' seja uma expressão válida:
expressao_dados <- "2.V002/1.V001" # Exemplo de expressão: média de moradores por domicílio
mapa <- mapa_da_cidade(cidade = "VITÓRIA", dados = expressao_dados)

}
\seealso{
Para obter detalhes sobre as variáveis disponíveis para uso na expressão de dados, consulte o data.frame \code{dicionario}.
}
