% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dados.R
\encoding{UTF-8}
\name{obter_dados}
\alias{obter_dados}
\title{Obter Dados com Base em uma Expressão}
\usage{
obter_dados(expressao)
}
\arguments{
\item{expressao}{Uma expressão que referencia variáveis nos dados do censo usando a notação "tabela.variavel".}
}
\value{
Um `data.frame` contendo os resultados da expressão fornecida.
}
\description{
Esta função avalia uma expressão fornecida e retorna um `data.frame` com os resultados.
A expressão deve referenciar variáveis disponíveis nos dados do censo, usando a notação "tabela.variavel".
}
\details{
As variáveis que podem ser referenciadas na expressão, bem como detalhes sobre elas,
estão disponíveis no data.frame "dicionario" incluído no pacote. Este data.frame fornece
uma descrição detalhada de cada variável disponível.
}
\examples{
# Suponha que expressao seja uma expressão válida que referencia variáveis nos dados do censo:
# obter_dados(expressao)

}
