% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tabelas}
\alias{tabelas}
\title{Tabelas de Dados}
\format{
Um data frame com 3 linhas e 2 colunas:
\describe{
  \item{tabela}{Numérico. Identificador da tabela. Exemplo: 1.}
  \item{Nome}{Caracter. Nome do arquivo da tabela. Exemplo: "Basico_ES.csv".}
}
}
\source{
IBGE, Censo 2010.
}
\usage{
tabelas
}
\description{
Este data frame contém os nomes das tabelas de dados disponíveis.
}
